Completed Exercise 22-1
I used GIMP to try to do what the book explained, but they give examples
for Photoshop. In the syllabus, it says to NOT use Photoshop (which
isn't an issue anyway, since I don't own it), but it is a bit confusing.
I looked up a few guides on how to do what it was asking in terms of
reducing the file size, so after disabling dithering, using a
web-optimized color palette, exporting as GIF, etc, I hope the files are
OK. I did read the concepts they'd like us to apply in the book.