Completed Exercise 22-2
Once again, GIMP was my tool of choice for experimenting with what
options I had to compress these .psd and .tif files into .jpgs. It's a
handy utility, and by thoroughly familiarizing myself with the concepts
in the book, I hope this will be suitable. I can't help but wonder,
though - with the advent of LTE networks (both in speed and ubiquity),
how much of a concern realistically is the difference between an 80KB
and 40KB file? Maybe my T-Mobile unlimited data plan has spolied me, but
when it comes to data-intensive, rich-looking sites, I say bring it on.
Anyhow, GIMP was my tool, and here is the product of my toil. Hope it
hits the mark.